port eXide function doc lookup fixes
- allow lookup of functions in mapped modules (e.g., kwic)
- fix math module lookup problems
- add support for fn:-prefixed function lookup, respecting the prefix if supplied
- add comments to explain handling of various cases
- make variable names a little more verbose to ease maintenance

In addition, this commit:
- suppresses functions and variables marked as private

I tested cases with and without module imports and confirmed the autocomplete works in both scenarios. The only thing I wasn’t able to test was the “signature” parameter - which I couldn’t figure out how to trigger. @wolfgang, if you could tell me how to trigger a query involving the signature parameter, I can test this.

I think a future refactoring could probably unify the get-built-in-functions and get-imported-functions - but I thought I’d start with this first.